+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -176,7 +176,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KRDPJJRT7TXSD648P2R5A4NE"),
+  _id: z.string().default("01KRDY30KZAC68HTXSKNBBM3G5"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,